This change is necessary since the file watcher is broken for those import paths that differ from the default import path. In other words, the current implementation of the file watcher doesn't consider additional import paths provided by the `Options.path?` interface member (though the loader does!).
This change implements an algorithm that checks for file existence at
the default import path. If a file does not exist at the default import
path, provided additional import paths are then considered. If one is
found, the path is updated for the watcher. The fallback and default
filepath used is always the default import path.